refactor(back-end): move user data route & fix conditions route error handling
