+++ server/src/index.ts
@@ -6,16 +6,23 @@ import chalk from 'chalk'
 import compression from 'compression'
 
 import db from './services/db'
-import data from './routes/data'
 import { checkLogin, sessionStore, corsHandler, notFound } from './middleware'
 
 import routes from './routes'
+
+/**
+ * @deprecated
+ */
+import oldSessions from './routes/old-sessions'
 
 require('dotenv').config()
 
 const mongoClient = db.init(process.env.DB_NAME as string)
 
 express()
+  /**
+   * Middleware
+   */
   .use(helmet())
   .use(bodyParser.urlencoded({ extended: true }))
   .use(sessionStore(mongoClient))
@@ -25,12 +32,18 @@ express()
 
   .use(checkLogin)
 
+  /**
+   * Routes
+   */
   .use(routes)
+  .post('/old-sessions', oldSessions.submit)
 
-  .post('/old-sessions', data.oldSessions)
-
-  .get('/user', data.user)
-
+  /**
+   * Route not found middleware
+   */
   .use(notFound)
 
+  /**
+   * Start the server
+   */
   .listen(process.env.PORT, () => chalk.green(`[server] listening on port ${process.env.PORT}`))
