+++ server/src/routes/user-data/user-data.routes.ts
@@ -5,3 +5,4 @@ export default Router()
   .post('/gear', controller.updateGear)
   .post('/spots', controller.updateSpots)
   .post('/threshold', controller.updateThreshold)
+  .get('/user', controller.getUser)
