+++ server/src/routes/user-data/user-data.service.ts
@@ -1,8 +1,7 @@
 import db from '../../services/db'
 import spotData from '../../services/spot'
 
-import { Gear } from '../../../../shared/interfaces/Gear'
-import { Spot } from '../../../../shared/interfaces/Spot'
+import { Gear, Spot, User, Session } from '../../../../shared/interfaces'
 
 async function setGear (username: string, gear: Gear): Promise<void> {
   try {
@@ -44,8 +43,39 @@ async function setThreshold (username: string, threshold: number): Promise<void>
   }
 }
 
+async function getUser (username: string): Promise<User> {
+  try {
+    const user: User = await getUserData(username)
+
+    if (!user) throw new Error('no user')
+
+    return user
+  } catch (err) {
+    throw err
+  }
+}
+
+async function getUserData (name: string): Promise<User> {
+  try {
+    const userData: User[] = await db.get({ name })
+    const user = userData[0]
+
+    user.sessions = user.sessions!.sort((a, b) => {
+      return (new Date(a.date) as any) - (new Date(b.date) as any)
+    })
+
+    delete user._id
+    delete user.password
+
+    return user
+  } catch (err) {
+    throw err
+  }
+}
+
 export default {
   setGear,
   setSpots,
-  setThreshold
+  setThreshold,
+  getUser
 }
